mid Final

+++ BankTransections/src/p2/Transaction.java
@@ -2,7 +2,7 @@ package p2;
 
 public interface Transaction {
 	public void Withdraw(int amt);
-	public void Payment(int amt);
+	public void Payment(float amt);
 	public void Deposite(int amt);
 
 }
